RavenDB-24723 address PR comments
