+++ test/SlowTests/Server/Documents/AI/AiAgent/AiAgentClientApiBasics.cs
@@ -86,20 +86,20 @@ public class AiAgentClientApiBasics : RavenTestBase
 
         chat.SetUserPrompt("what goes well with my cheese?");
         var r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
         
         Assert.NotNull(chat.Answer);
         Assert.NotNull(chat.Id);
 
         chat.SetUserPrompt("what goes well with my cheese?");
         r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
 
         Assert.NotNull(chat.Answer);
 
         chat.SetUserPrompt("what cheese goes well with italian food?");
         r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
     }
 
     [RavenTheory(RavenTestCategory.Ai)]
@@ -150,7 +150,7 @@ public class AiAgentClientApiBasics : RavenTestBase
 
         var r = await chat.RunAsync(CancellationToken.None);
 
-        Assert.Equal(ConversationResult.ActionRequired, r);
+        Assert.Equal(AiConversationResult.ActionRequired, r);
         Assert.NotNull(chat.Id);
 
         foreach (var request in chat.RequiredActions())
@@ -160,7 +160,7 @@ public class AiAgentClientApiBasics : RavenTestBase
        
         r = await chat.RunAsync(CancellationToken.None);
 
-        if (r == ConversationResult.ActionRequired)
+        if (r == AiConversationResult.ActionRequired)
         {
             // agent could ask for action tools again
             Assert.True(chat.RequiredActions().Any());
@@ -225,13 +225,13 @@ public class AiAgentClientApiBasics : RavenTestBase
 
         chat.SetUserPrompt("what goes well with my cheese?");
         var r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
         
         Assert.NotNull(chat.Answer);
         Assert.NotNull(chat.Id);
 
         r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
         Assert.Equal(0, chat.RequiredActions().ToList().Count);
 
         chat.AddActionResponse("foo","bar");
@@ -245,7 +245,7 @@ public class AiAgentClientApiBasics : RavenTestBase
 
         chat.SetUserPrompt("what cheese goes well with italian food?");
         r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
     }
 
     [RavenTheory(RavenTestCategory.Ai)]
@@ -289,7 +289,7 @@ public class AiAgentClientApiBasics : RavenTestBase
             p => p.AddParameter("company", "companies/90-A"));
         chat.SetUserPrompt("what goes well with my cheese?");
         var r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
 
         chat = store.AI.ResumeConversation<OutputSchema>(chat.Id, "foo");
         chat.SetUserPrompt("Can you give me some alternatives?");
@@ -298,12 +298,12 @@ public class AiAgentClientApiBasics : RavenTestBase
 
         chat.SetUserPrompt("Can you give me some alternatives?");
         r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
         Assert.NotNull(chat.ChangeVector);
 
         chat.SetUserPrompt("even better choice?");
         r = await chat.RunAsync(CancellationToken.None);
-        Assert.Equal(ConversationResult.Done, r);
+        Assert.Equal(AiConversationResult.Done, r);
         Assert.NotNull(chat.ChangeVector);
     }
 }
